+++ modules/nested-tabs-html/assets/scss/frontend.scss
@@ -3,6 +3,46 @@
 //
 // Nested Tabs
 //
+
+@mixin active-tab-title() {
+
+	&,
+	& a {
+		color: var( --n-tabs-title-color-active );
+	}
+
+	.e-n-tab-icon {
+
+		i {
+			color: var( --n-tabs-icon-color-active );
+		}
+
+		svg {
+			fill: var( --n-tabs-icon-color-active );
+		}
+
+		i,
+		svg {
+
+			&:first-child {
+				transform: translate(0, -100vh);
+				height: 0;
+				opacity: 0;
+			}
+		}
+	}
+
+	// Remove the hover animation from the active tabs.
+	&[class*="elementor-animation-"] {
+		&:focus,
+		&:active,
+		&:hover {
+			transform: initial;
+			animation: initial;
+		}
+	}
+}
+
 .elementor-widget-n-tabs-html {
 	--n-tabs-color-accent-fallback: #61CE70;
 	--n-tabs-color-secondary-fallback: #54595F;
@@ -31,6 +71,7 @@
 	--n-tabs-title-width: initial;
 	--n-tabs-title-height: initial;
 	--n-tabs-title-font-size: 1rem;
+	--n-tabs-title-white-space: initial;
 	--n-tabs-title-justify-content-toggle: initial;
 	--n-tabs-title-align-items-toggle: center;
 	--n-tabs-title-justify-content: center;
@@ -117,7 +158,7 @@
 
 		&-title {
 			all: unset;
-			white-space: nowrap;
+			user-select: none;
 			display: flex;
 			align-items: var( --n-tabs-title-align-items-toggle, var( --n-tabs-title-align-items ) );
 			flex-direction: var( --n-tabs-title-direction );
@@ -136,25 +177,17 @@
 			border-radius: var( --n-tabs-title-border-radius );
 			height: var( --n-tabs-title-height );
 			width: var( --n-tabs-title-width );
+			white-space: var( --n-tabs-title-white-space );
 			transition: background var( --n-tabs-title-transition),
-			color var( --n-tabs-title-transition),
-			border var( --n-tabs-title-transition),
-			box-shadow var( --n-tabs-title-transition),
-			text-shadow var( --n-tabs-title-transition),
-			stroke var( --n-tabs-title-transition),
-			stroke-width var( --n-tabs-title-transition),
-			-webkit-text-stroke-width var( --n-tabs-title-transition),
-			-webkit-text-stroke-color var( --n-tabs-title-transition),
-			transform var( --n-tabs-title-transition );
-
-			// Remove tap highlight color on iOS: https://developer.mozilla.org/en-US/docs/Web/CSS/-webkit-tap-highlight-color.
-			-webkit-tap-highlight-color: transparent;
-			-webkit-touch-callout: none;
-			-webkit-user-select: none;
-			-khtml-user-select: none;
-			-moz-user-select: none;
-			-ms-user-select: none;
-			user-select: none;
+				color var( --n-tabs-title-transition),
+				border var( --n-tabs-title-transition),
+				box-shadow var( --n-tabs-title-transition),
+				text-shadow var( --n-tabs-title-transition),
+				stroke var( --n-tabs-title-transition),
+				stroke-width var( --n-tabs-title-transition),
+				-webkit-text-stroke-width var( --n-tabs-title-transition),
+				-webkit-text-stroke-color var( --n-tabs-title-transition),
+				transform var( --n-tabs-title-transition );
 
 			@include button-focus;
 
@@ -222,47 +255,12 @@
 			}
 
 			&[aria-selected="true"] {
-
-				&,
-				& a {
-					color: var( --n-tabs-title-color-active );
-				}
-
-				.e-n-tab-icon {
-
-					i {
-						color: var( --n-tabs-icon-color-active );
-					}
-
-					svg {
-						fill: var( --n-tabs-icon-color-active );
-					}
-
-					i,
-					svg {
-
-						&:first-child {
-							transform: translate(0, -100vh);
-							height: 0;
-							opacity: 0;
-						}
-					}
-				}
-
-				// Remove the hover animation from the active tabs.
-				&[class*="elementor-animation-"] {
-					&:focus,
-					&:active,
-					&:hover {
-						transform: initial;
-						animation: initial;
-					}
-				}
+				@include active-tab-title;
 			}
 		}
 	}
 
-	[data-layout="tabs"] {
+	[data-touch-mode="false"] {
 
 		.e-n-tab-title[aria-selected="false"]:hover {
 
@@ -283,9 +281,17 @@
 			}
 		}
 	}
+
+	[data-touch-mode="true"] {
+
+		.e-n-tab-title[aria-selected="false"]:hover {
+
+			@include active-tab-title;
+		}
+	}
 }
 
-.elementor .elementor-element.elementor-widget-n-tabs-html > .elementor-widget-container > .e-n-tabs[data-layout="tabs"] > .e-n-tabs-heading .e-n-tab-title[aria-selected="false"]:hover {
+.elementor .elementor-element.elementor-widget-n-tabs-html > .elementor-widget-container > .e-n-tabs[data-touch-mode="false"] > .e-n-tabs-heading .e-n-tab-title[aria-selected="false"]:hover {
 	background-color: var( --n-tabs-title-background-color-hover );
 	background-image: initial;
 }
