+++ tests/playwright/sanity/modules/nested-tabs-html/nested-tabs-html.test.ts-snapshots/tabs-direction-top-icon-position-top-align-default-linux.jpeg
